backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
sqlalchemy==2.0.23
asyncpg==0.29.0
qdrant-client==1.7.0
openai==1.3.8
python-multipart==0.0.6
better-exceptions==0.3.3
python-dotenv==1.0.0
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-json-logger==2.0.7
sentence-transformers==2.2.2